migrate top level files to glintwebui to further modularize glint
