+++ glintwebui/glint_api.py
@@ -2,7 +2,7 @@ from keystoneclient.auth.identity import v2, v3
 from keystoneauth1 import session
 from keystoneauth1 import exceptions
 import glanceclient
-import glintv2.config as config
+import .config as config
 import json
 import logging
 import os
