+++ glintwebui/urls.py
@@ -2,6 +2,11 @@ from django.conf.urls import url
 
 from . import views
 
+
+from .celery_app import image_collection
+from .utils import check_collection_task, set_collection_task
+
+
 urlpatterns = [
     url(r'^$', views.index, name='index'),
     url(r'^project_details/(?P<account_name>.+)/$', views.project_details, name='project_details'),
@@ -25,3 +30,9 @@ urlpatterns = [
     url(r'^upload_image/(?P<account_name>.+)/$', views.upload_image, name='upload_image'),
     url(r'^save_hidden_images/(?P<account_name>.+)/$', views.save_hidden_images, name='save_hidden_images'),
 ]
+
+# Check if the image collection task is running, if not start it and set it to running
+collection_started = check_collection_task()
+if not collection_started:
+    image_collection.apply_async(queue='image_collection')
+    set_collection_task(True)
